+++ lib/sidekiq/launcher.rb
@@ -96,6 +96,24 @@ module Sidekiq
       ❤
     end
 
+    def self.flush_stats
+      nowdate = Time.now.utc.strftime("%Y-%m-%d")
+      fails = Processor::FAILURE.reset
+      procd = Processor::PROCESSED.reset
+      Sidekiq.redis do |conn|
+        conn.pipelined do
+          conn.incrby("stat:processed", procd)
+          conn.incrby("stat:processed:#{nowdate}", procd)
+          conn.expire("stat:processed:#{nowdate}", STATS_TTL)
+
+          conn.incrby("stat:failed", fails)
+          conn.incrby("stat:failed:#{nowdate}", fails)
+          conn.expire("stat:failed:#{nowdate}", STATS_TTL)
+        end
+      end
+    end
+    at_exit(&method(:flush_stats))
+
     def ❤
       key = identity
       fails = procd = 0